Refactored module organization now have transacting.py and devicing.py split out of stacking.py
